requirements.txt
streamlit
pandas
altair
pillow
numpy
google-genai
python-dotenv
scikit-image
mcrcon
plotly
